Move component to folder components of module
