+++ frontend/src/app/modules/search-cards/components/search-cards-header/search-cards-header.component.html
@@ -1,0 +1,1 @@
+<p>search-cards-header works!</p>
