+++ frontend/src/app/modules/search-cards/components/search-cards-header/search-cards-header.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-search-cards-header',
+  templateUrl: './search-cards-header.component.html',
+  styleUrl: './search-cards-header.component.scss'
+})
+export class SearchCardsHeaderComponent {
+
+}
