+++ frontend/src/app/modules/search-cards/components/search-cards-main/search-cards-main.component.spec.ts
